# internal/api/gophermart-types.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.1.0 DO NOT EDIT.
package api

import (
	"time"
)

// Balance defines model for Balance.
type Balance struct {
	Current   float64 `json:"current"`
	Withdrawn float64 `json:"withdrawn"`
}

// Error defines model for Error.
type Error struct {
	// Code Error code
	Code int `json:"code"`

	// Message Error message
	Message string `json:"message"`
}

// Order defines model for Order.
type Order struct {
	Accrual    *float64   `json:"accrual,omitempty"`
	Number     *string    `json:"number,omitempty"`
	Status     *string    `json:"status,omitempty"`
	UploadedAt *time.Time `json:"uploaded_at,omitempty"`
}

// User defines model for User.
type User struct {
	Login    string `json:"login"`
	Password string `json:"password"`
}

// Withdraw defines model for Withdraw.
type Withdraw struct {
	Order       string     `json:"order"`
	ProcessedAt *time.Time `json:"processed_at,omitempty"`
	Sum         float64    `json:"sum"`
}

// UploadOrderTextBody defines parameters for UploadOrder.
type UploadOrderTextBody = string

// WithdrawRequestJSONRequestBody defines body for WithdrawRequest for application/json ContentType.
type WithdrawRequestJSONRequestBody = Withdraw

// LoginUserJSONRequestBody defines body for LoginUser for application/json ContentType.
type LoginUserJSONRequestBody = User

// UploadOrderTextRequestBody defines body for UploadOrder for text/plain ContentType.
type UploadOrderTextRequestBody = UploadOrderTextBody

// RegisterUserJSONRequestBody defines body for RegisterUser for application/json ContentType.
type RegisterUserJSONRequestBody = User